# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "MyWebScrcpy API", "version": "1.6.0", "description": "当前服务注册的 HTTP 与 WebSocket 接口。服务默认无内建身份认证，请在受信网络中使用。"},
  "servers": [{"url": "/"}],
  "paths": {
...
    "/api/rotate":{"get":{"tags":["设备与控制"],"summary":"切换设备旋转方向","parameters":[{"$ref":"#/components/parameters/serial"}],"x-side-effect":true,"responses":{"200":{"description":"旋转结果"},"400":{"description":"缺少 serial"}}}},
    "/api/screen-state":{"get":{"tags":["设备与控制"],"summary":"读取屏幕电源状态","parameters":[{"$ref":"#/components/parameters/serial"}],"responses":{"200":{"description":"screenOn"}}}},
    "/api/reboot":{"post":{"tags":["设备与控制"],"summary":"重启设备","parameters":[{"$ref":"#/components/parameters/serial"}],"x-side-effect":true,"responses":{"200":{"description":"成功"},"502":{"description":"ADB 失败"}}}},
...
    "/api/ui/xml":{"get":{"tags":["UI 检查"],"summary":"读取 Android UI XML","parameters":[{"$ref":"#/components/parameters/serial"}]}},
    "/api/ui/page-info":{"get":{"tags":["UI 检查"],"summary":"读取前台应用与屏幕信息","parameters":[{"$ref":"#/components/parameters/serial"}]}},
    "/api/vision/stream":{"get":{"tags":["Vision"],"summary":"Vision WebSocket 双向帧/动作通道","description":"建立与目标设备的 Vision 双向 WebSocket 通道。连接后服务端先发送 hello（含 protocol_version、device_id、session_id、width、height、codec），随后持续推送二进制视频帧；客户端通过文本 JSON 上行消息。\n\n上行文本消息（客户端→服务端）：\n- hello：携带 max_fps（可选），服务端回 hello.ack。\n- action.request：请求执行设备动作（tap/swipe/key/text），字段定义见 components.schemas.VisionActionRequest，其中 humanize=true 启用拟人化触摸。\n- detection.result：回传识别结果，服务端校验后广播。\n- heartbeat：保活，服务端回 heartbeat。\n\n下行文本消息（服务端→客户端）：\n- hello / hello.ack / action.result / heartbeat / error。\n\naction.request 返回 action.result：{\"type\":\"action.result\",\"request_id\":\"...\",\"accepted\":true,\"executed\":true,\"error_code\":\"...\"}。\n\n动作会与浏览器控制消息共用同一设备级队列；坐标越界返回 invalid_coordinate，队列满返回 queue_full，超时返回 expired。","parameters":[{"$ref":"#/components/parameters/serial"}]}},
...
  },
  "components":{
    "parameters":{"serial":{"name":"serial","in":"query","required":true,"schema":{"type":"string"},"description":"ADB 设备 serial"}},
...
  }
}